requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-localflavor==4.0
django-precise-bbcode==1.2.16
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.1
pillow==10.2.0
python-stdnum==1.20
sqlparse==0.4.4
tzdata==2023.4
